+++ DarkCodex/MartialArt.cs
@@ -45,14 +45,97 @@ namespace DarkCodex
                 "At 3rd level, a virtuous bravo gains a +1 dodge bonus to AC while wearing light armor or no armor.\nAnything that causes the virtuous bravo to lose her Dexterity bonus to AC also causes her to lose this dodge bonus. This bonus increases by 1 for every 4 paladin levels beyond 3rd (to a maximum of +5 at 19th level).\nThis ability replaces mercy."
                 ).SetComponents(
                 new CreateAddStatBonusInArmor(Resource.ValueRank, StatType.AC, ModifierDescriptor.Dodge, ArmorProficiencyGroup.None, ArmorProficiencyGroup.Light),
-                Helper.CreateContextRankConfig(ContextRankBaseValueType.ClassLevel, ContextRankProgression.StartPlusDivStep, AbilityRankType.Default, 
+                Helper.CreateContextRankConfig(ContextRankBaseValueType.ClassLevel, ContextRankProgression.StartPlusDivStep, AbilityRankType.Default,
                     startLevel: 3, stepLevel: 4, classes: paladin.ObjToArray())
+                );
+
+            var resourcePanache = Helper.CreateBlueprintAbilityResource(
+                "PanacheResource",
+                "Panache",
+                "A swashbuckler fights with panache: a fluctuating measure of a swashbuckler’s ability to perform amazing actions in combat. At the start of each day, a swashbuckler gains a number of panache points equal to her Charisma modifier (minimum 1). Her panache goes up or down throughout the day, but cannot go higher than her Charisma modifier (minimum 1). A swashbuckler spends panache to accomplish deeds, and regains panache in the following ways.\nCritical Hit with a Light or One-Handed Piercing Melee Weapon: Each time the swashbuckler confirms a critical hit with a light or one-handed piercing melee weapon, she regains 1 panache point. Confirming a critical hit on a helpless or unaware creature or a creature that has fewer Hit Dice than half the swashbuckler’s character level doesn’t restore panache.\nKilling Blow with a Light or One-Handed Piercing Melee Weapon: When the swashbuckler reduces a creature to 0 or fewer hit points with a light or one-handed piercing melee weapon attack while in combat, she regains 1 panache point. Destroying an unattended object, reducing a helpless or unaware creature to 0 or fewer hit points, or reducing a creature that has fewer Hit Dice than half the swashbuckler’s character level to 0 or fewer hit points doesn’t restore any panache.",
+                stat: StatType.Charisma
+                ).ToReference<BlueprintAbilityResourceReference>();
+
+            var panache1 = Helper.CreateBlueprintFeature(
+                "Panache_PreciseStrike",
+                "Precise Strike",
+                "At 3rd level, while she has at least 1 panache point, a swashbuckler gains the ability to strike precisely with a light or one-handed piercing melee weapon (though not natural weapon attacks), adding her swashbuckler level to the damage dealt. To use this deed, a swashbuckler cannot attack with a weapon in her other hand or use a shield other than a buckler. She can even use this ability with thrown light or one-handed piercing melee weapons, so long as the target is within 30 feet of her. Any creature that is immune to sneak attacks is immune to the additional damage granted by precise strike, and any item or ability that protects a creature from critical hits also protects a creature from the additional damage of a precise strike. This additional damage is precision damage, and isn’t multiplied on a critical hit."
+                ).SetComponents(
+                new DuelistPreciseStrike { m_Duelist = paladin }
+                );
+
+            var panache2 = Helper.CreateBlueprintFeature(
+                "Panache_SwashbucklerInitiative",
+                "Swashbuckler Initiative",
+                "At 3rd level, while the swashbuckler has at least 1 panache point, she gains a +2 bonus on initiative checks. In addition, if she has the Quick Draw feat, her hands are free and unrestrained, and she has any single light or one-handed piercing melee weapon that isn’t hidden, she can draw that weapon as part of the initiative check."
+                ).SetComponents(
+                Helper.CreateAddStatBonus(2, StatType.Initiative)
+                );
+
+            var panache3 = Helper.CreateBlueprintFeature(
+                "Panache_DodgingPanache",
+                "Dodging Panache",
+                "At 1st level, when an opponent attempts a melee attack against the swashbuckler, the swashbuckler can as an immediate action spend 1 panache point to gain a dodge bonus to AC equal to her Charisma modifier (minimum 0) against the triggering attack. The swashbuckler can only perform this deed while wearing light or no armor, and while carrying no heavier than a light load."
+                ).SetComponents(
+                new PanacheDodge(resourcePanache)
+                );
+
+            var panache4 = Helper.CreateBlueprintFeature(
+                "Panache_OpportuneParryRiposte",
+                "Opportune Parry and Riposte",
+                "At 1st level, the swashbuckler gains the duelist’s parry class feature, but may only parry attacks targeted at herself. Upon performing a successful parry and if she has at least 1 panache point, the swashbuckler can as an immediate action make an attack against the creature whose attack she parried, provided that creature is within her reach."
+                ).SetComponents(
+                Helper.CreateAddFacts("c0248f304998aa64da458e097c022d73"), //DuelistParrySelfAbility
+                Helper.CreateAddMechanicsFeature(AddMechanicsFeature.MechanicsFeatureType.DuelistRiposte)
+                );
+
+            var swordplay_ab = Helper.CreateBlueprintActivatableAbility(
+                "MenacingSwordplayActivatable",
+                out var buff,
+                "Menacing Swordplay",
+                "At 3rd level, while she has at least 1 panache point, when a swashbuckler hits an opponent with a light or one-handed piercing melee weapon, she can choose to use Intimidate to demoralize that opponent as a swift action instead of a standard action.",
+                icon: null,
+                onByDefault: true,
+                deactivateWhenStunned: true,
+                activationType: AbilityActivationType.WithUnitCommand,
+                commandType: UnitCommand.CommandType.Swift
+                ).SetComponents(
+                new DeactivateImmediatelyIfNoAttacksThisRound(),
+                new ActivatableAbilityUnitCommand { Type = UnitCommand.CommandType.Swift }
+                );
+            buff.SetComponents(
+                Helper.CreateAddInitiatorAttackWithWeaponTrigger(
+                    Helper.Get<BlueprintAbility>("7d2233c3b7a0b984ba058a83b736e6ac").GetComponent<AbilityEffectRunAction>().Actions, //PersuasionUseAbility
+                    DuelistWeapon: true));
+            var panache5 = Helper.CreateBlueprintFeature(
+                "Panache_MenacingSwordplay"
+                ).SetComponents(
+                Helper.CreateAddFacts(swordplay_ab)
+                ).SetUIData(swordplay_ab);
+            
+            var f4_panache = Helper.CreateBlueprintFeature(
+                "VirtuousBravoPanacheDeeds",
+                "Panache and Deeds",
+                "At 4th level, a virtuous bravo gains the swashbuckler’s panache class feature along with the following swashbuckler deeds: dodging panache, menacing swordplay, opportune parry and riposte, precise strike, and swashbuckler initiative. The virtuous bravo’s paladin levels stack with any swashbuckler levels when using these deeds.\nThis ability replaces the paladin’s spellcasting."
+                ).SetComponents(
+                Helper.CreateAddAbilityResources(resourcePanache),
+                Helper.CreateAddFacts(panache1, panache2, panache3, panache4, panache5),
+                Helper.CreateAddInitiatorAttackWithWeaponTrigger(
+                    Helper.CreateActionList(new ContextRestoreResource { m_Resource = resourcePanache }),
+                    ActionsOnInitiator: true,
+                    DuelistWeapon: true));
+
+            var f11_panache = Helper.CreateBlueprintFeature(
+                "VirtuousBravoAdvancedDeeds",
+                "Advanced Deeds",
+                "At 11th level, a virtuous bravo gains the following swashbuckler deeds: bleeding wound, evasive, subtle blade, superior feint, swashbuckler’s grace, and targeted strike.\nThis ability replaces aura of justice."
                 );
 
             archetype.SetAddFeatures(
                 Helper.CreateLevelEntry(1, f1_prof),
                 Helper.CreateLevelEntry(1, f1_finesse),
-                Helper.CreateLevelEntry(3, f3_nimble)
+                Helper.CreateLevelEntry(3, f3_nimble),
+                Helper.CreateLevelEntry(4, f4_panache)
                 );
 
             archetype.SetRemoveFeatures(
@@ -74,20 +157,6 @@ namespace DarkCodex
         As a move action, you can shorten your grip on the glaive, treating it as though it lacked the reach weapon property. You can adjust your grip to grant the weapon the reach property as a move action.
         
         Virtuous Bravo
-
-        Panache and Deeds (Ex)
-        At 4th level, a virtuous bravo gains the swashbuckler’s panache class feature along with the following swashbuckler deeds: dodging panache, menacing swordplay, opportune parry and riposte, precise strike, and swashbuckler initiative. The virtuous bravo’s paladin levels stack with any swashbuckler levels when using these deeds.
-        This ability replaces the paladin’s spellcasting.
-
-        *Dodging Panache (Ex): At 1st level, when an opponent attempts a melee attack against the swashbuckler, the swashbuckler can as an immediate action spend 1 panache point to move 5 feet; doing so grants the swashbuckler a dodge bonus to AC equal to her Charisma modifier (minimum 0) against the triggering attack. This movement doesn’t negate the attack, which is still resolved as if the swashbuckler had not moved from the original square. This movement is not a 5-foot step; it provokes attacks of opportunity from creatures other than the one who triggered this deed. The swashbuckler can only perform this deed while wearing light or no armor, and while carrying no heavier than a light load.
-        *Menacing Swordplay (Ex): At 3rd level, while she has at least 1 panache point, when a swashbuckler hits an opponent with a light or one-handed piercing melee weapon, she can choose to use Intimidate to demoralize that opponent as a swift action instead of a standard action.
-        *Opportune Parry and Riposte (Ex): At 1st level, when an opponent makes a melee attack against the swashbuckler, she can spend 1 panache point and expend a use of an attack of opportunity to attempt to parry that attack. The swashbuckler makes an attack roll as if she were making an attack of opportunity; for each size category the attacking creature is larger than the swashbuckler, the swashbuckler takes a –2 penalty on this roll. If her result is greater than the attacking creature’s result, the creature’s attack automatically misses. The swashbuckler must declare the use of this ability after the creature’s attack is announced, but before its attack roll is made. Upon performing a successful parry and if she has at least 1 panache point, the swashbuckler can as an immediate action make an attack against the creature whose attack she parried, provided that creature is within her reach. This deed’s cost cannot be reduced by any ability or effect that reduces the number of panache points a deed costs.
-        *Precise Strike (Ex): At 3rd level, while she has at least 1 panache point, a swashbuckler gains the ability to strike precisely with a light or one-handed piercing melee weapon (though not natural weapon attacks), adding her swashbuckler level to the damage dealt. To use this deed, a swashbuckler cannot attack with a weapon in her other hand or use a shield other than a buckler. She can even use this ability with thrown light or one-handed piercing melee weapons, so long as the target is within 30 feet of her. Any creature that is immune to sneak attacks is immune to the additional damage granted by precise strike, and any item or ability that protects a creature from critical hits also protects a creature from the additional damage of a precise strike. This additional damage is precision damage, and isn’t multiplied on a critical hit. As a swift action, a swashbuckler can spend 1 panache point to double her precise strike’s damage bonus on the next attack. This benefit must be used before the end of her turn, or it is lost. This deed’s cost cannot be reduced by any ability or effect that reduces the amount of panache points a deed costs (such as the Signature Deed feat).
-        *Swashbuckler Initiative (Ex): At 3rd level, while the swashbuckler has at least 1 panache point, she gains a +2 bonus on initiative checks. In addition, if she has the Quick Draw feat, her hands are free and unrestrained, and she has any single light or one-handed piercing melee weapon that isn’t hidden, she can draw that weapon as part of the initiative check.
-        
-        Advanced Deeds (Ex)
-        At 11th level, a virtuous bravo gains the following swashbuckler deeds: bleeding wound, evasive, subtle blade, superior feint, swashbuckler’s grace, and targeted strike.
-        This ability replaces aura of justice.
 
         *Bleeding Wound (Ex): At 11th level, when the swashbuckler hits a living creature with a light or one-handed piercing melee weapon attack, as a free action she can spend 1 panache point to have that attack deal additional bleed damage. The amount of bleed damage dealt is equal to the swashbuckler’s Dexterity modifier (minimum 1). Alternatively, the swashbuckler can spend 2 panache points to deal 1 point of Strength, Dexterity, or Constitution bleed damage instead (swashbuckler’s choice). Creatures that are immune to sneak attacks are also immune to these types of bleed damage.
         *Evasive (Ex): At 11th level, while a swashbuckler has at least 1 panache point, she gains the benefits of the evasion, uncanny dodge, and improved uncanny dodge rogue class features. She uses her swashbuckler level as her rogue level for improved uncanny dodge.